// src/app/types/Restypes.tsx
export interface User {
  id: number;
  second_id: number;
  first_name: string;
  last_name: string;
  profile_pic: string;
  last_message: string;
  last_seen: string;
  is_online: number; // 1 for online, 0 for offline
  unread_message: number;
  user_role: string;
  time: string; // Last active time
  is_block: number; // 1 for blocked, 0 for not blocked
}